Game Over and Grinder
Gave grinder prefab a hitbox for new hazard mechanic. When player runs out of health or hits a hazard (grinder, for example), the level is reset. When an object hits a hazard, the object is respawned at its initial spawn point. When an enemy hits a hazard, they are removed from the level.